+++ components/BadgeDesign.module.scss
